Added Accumulator engine class; extended Redactable Signature engine class.
Started implementing Pöhls' and Samelin's Redactable Signature Scheme. The
Accumulator and RedactableSignature SPIs for this scheme are still empty
skeletons. Only the KeyPairGenerator is fully implemented at this point. The
PSRSS scheme provides update and merge operations, so the RedactableSignature
engine class had to be adjusted. The scheme uses an accumulator, which did not
have a engine class.