Implemented algorithm independent instantiation of ModificationInstruction.
ModificationInstruction now has a static factory method that creates a new
ModificationInstruction for the given algorithm. The only (abstract) public
method of ModificationInstruction is `add(byte[] part)`. This allows
implementors to define own methods to retrieve parts from the
ModificationInstruction. And Note that every part of a message can be
identified by an byte array, because parts are represented as byte[].