Implement sign algorithm of general list based RSS
Implemented the sign algorithm for linear documents as described in the
"General Framework..." paper. This algorithm is based on the general
RSS for sets. The general RSS for sets is renamed to GSRSS and the
general RSS for lists is named GLRSS.

Also renamed the ByteArrayWrapper class to ByteArray for brevity.