Make RedactableSignature support list based RSS
The Identifier class was added to support list and set based RSS
implementations. The identifier is generated by the underlying
implementation when RedactableSignature.addPart() is called. The
Identifier is later used to specify the elements to redact.

This commit also changes how the RedactableSignature engine class is
used. Since the SignatureOutput is "standalone", there is no need to
call the `addPart()` method (same is true for mergeing). The new
behaviour is the following:

Sign:
    1. void init(KeyPair)
    2. Identifier addPArt(byte[], boolean)
    3. SignatureOutput sign()

Verify:
    1. void init(PublicKey)
    2. boolean verify(SignatureOutput)

Redact:
    1. void init(PublicKey)
    2. void addIdentifier(Identifier)
    3. SignatureOutput redact(SignatureOutput)

Update:
    1. void init(KeyPair)
    2. Identifier addPart(byte[], boolean)
    3. SignatureOutput update(SignatureOutput)

Merge:
    1. void init(KeyPair)
    2. SignatureOutput merge(SignatureOutput, SignatureOutput)