Fix potential mix and match attack on GSRSS
Non redactable message parts were not protected by the accumulator,
which could allow a mix and match attack on the GSRSS implementation.
GSRSS is now implemented as described in the "General Framework..."
paper, which creates witnesses for ALL message parts. This fix also
allows a more streamlined implementation of the XML encoding of GSRSS.