Switch XML marshalling back to DOM
The previous approach using JAXB was not suitable for implementing the
SignatureInfo element. The SignatureInfo element has to be generated
before all of the XML is generated. JAXB does not allow to generate a
byte by byte identical subtree without the surrounding elements, which
made it impossible to create the SignatureInfo element before the rest
of the XML exists.